+++ pkg/generator/fixtures/module_outputs/.terraform/modules/1894f30c892e8475e20725b493c6ce4b/outputs.tf
@@ -1,0 +1,4 @@
+output "first_output" {
+  description = "First output"
+  value       = "${random_string.password.result}"
+}