+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## (Unreleased)
 
+NEW FEATURES:
+* Add generator for re-outputting modules outputs
+
 ENHANCEMENTS:
 * Optimize project structure, split to separate packages and add cobra for cli